backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
Pillow==10.1.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
google-cloud-vision==3.4.5
google-cloud-speech==2.21.0
twilio==8.10.0
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0